+++ src/Tracking/Request/UnsupportedTrackingPayloadException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Zhortein\SeoTrackingBundle\Tracking\Request;
+
+final class UnsupportedTrackingPayloadException extends \InvalidArgumentException
+{
+}